init (#703)
